@@ -2,21 +2,21 @@ apply plugin: 'com.android.library'
 
 def artifactGroupId = 'jp.realglobe'
 def artifactArtifactId = 'android-uploader'
-def artifactVersion = '3.2.0'
+def artifactVersion = '3.2.1'
 def artifactName = 'Uploader'
 def artifactDescription = 'Uploader'
 def artifactUrl = "https://github.com/realglobe-Inc/${artifactArtifactId}"
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 19
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName artifactVersion
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
     }
 
@@ -36,12 +36,12 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test:runner:1.2.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 
-    implementation 'jp.realglobe:android-lib:5.3.0'
+    implementation 'jp.realglobe:android-lib:6.0.2'
 }
 
 apply plugin: 'com.jfrog.bintray'
