@@ -19,8 +19,9 @@ package jp.realglobe.android.uploader;
 import android.os.Handler;
 import android.os.Looper;
 import android.os.Message;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import java.io.BufferedOutputStream;
 import java.io.OutputStream;
@@ -37,8 +38,6 @@ import jp.realglobe.android.function.Functions;
  * Created by fukuchidaisuke on 17/07/05.
  */
 public class Poster {
-
-    private static final String TAG = Poster.class.getName();
 
     private static final int DEFAULT_TIMEOUT = 30_000; // ミリ秒
 
@@ -174,11 +173,9 @@ public class Poster {
 
         @Override
         public void handleMessage(Message msg) {
-            switch (msg.what) {
-                case MSG_POST: {
-                    final Entry entry = (Entry) msg.obj;
-                    post(entry);
-                }
+            if (msg.what == MSG_POST) {
+                final Entry entry = (Entry) msg.obj;
+                post(entry);
             }
         }
 
