@@ -1,8 +1,9 @@
 package jp.realglobe.android.uploader;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.util.Base64;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
@@ -159,7 +160,7 @@ public class MultipartEntryBuilder implements Poster.EntryBuilder {
      */
     @NonNull
     public MultipartEntryBuilder setHeader(@Nullable Map<String, String> header) {
-        this.header = new HashMap<>(header);
+        this.header = header == null ? new HashMap<>() : new HashMap<>(header);
         return this;
     }
 
